+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [1.3.2] - 2026-03-04
 
 - CSS: Add chip
-- React components: Add FilterChips
+- React components: Add chip and FilterChips
 
 ## [1.3.1] - 2026-02-20